@@ -15,20 +15,26 @@
   "dependencies": {
     "@airship/with-unstated": "^1.0.1",
     "@react-native-community/async-storage": "^1.6.2",
+    "@react-native-community/masked-view": "0.1.6",
+    "@react-navigation/bottom-tabs": "^5.5.2",
+    "@react-navigation/drawer": "^5.7.7",
+    "@react-navigation/material-top-tabs": "^5.2.8",
+    "@react-navigation/native": "^5.4.2",
+    "@react-navigation/stack": "^5.3.9",
     "axios": "^0.18.1",
-    "expo": "^35.0.0",
-    "expo-av": "~7.0.0",
-    "expo-asset": "~7.0.0",
+    "expo": "^37.0.0",
+    "expo-asset": "~8.1.5",
+    "expo-av": "~8.1.0",
     "expo-cli": "^3.4.0",
-    "expo-constants": "~7.0.0",
-    "expo-file-system": "^8.0.0",
-    "expo-font": "~7.0.0",
-    "expo-image-manipulator": "^8.0.0",
-    "expo-image-picker": "^8.0.2",
-    "expo-localization": "~7.0.0",
-    "expo-location": "~7.0.0",
-    "expo-permissions": "~7.0.0",
-    "expo-web-browser": "~7.0.0",
+    "expo-constants": "~9.0.0",
+    "expo-file-system": "~8.1.0",
+    "expo-font": "~8.1.0",
+    "expo-image-manipulator": "~8.1.0",
+    "expo-image-picker": "~8.1.0",
+    "expo-localization": "~8.1.0",
+    "expo-location": "~8.1.0",
+    "expo-permissions": "~8.1.0",
+    "expo-web-browser": "~8.2.1",
     "fbemitter": "^2.1.1",
     "hoist-non-react-statics": "^2.3.1",
     "i18n-js": "^3.3.0",
@@ -36,12 +42,12 @@
     "moment": "^2.20.1",
     "moment-timezone": "^0.5.26",
     "ramda": "^0.25.0",
-    "react": "16.8.3",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
+    "react": "16.9.0",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-37.0.1.tar.gz",
     "react-native-animatable": "^1.3.2",
     "react-native-deck-swiper": "^1.6.7",
     "react-native-fade-in-image": "1.3.0",
-    "react-native-gesture-handler": "1.3.0",
+    "react-native-gesture-handler": "~1.6.0",
     "react-native-image-zoom-viewer": "^2.2.27",
     "react-native-material-menu": "^1.0.0",
     "react-native-mime-types": "^2.2.1",
@@ -49,10 +55,12 @@
     "react-native-parsed-text": "0.0.21",
     "react-native-popup-menu": "^0.15.6",
     "react-native-read-more-text": "^1.0.0",
-    "react-native-reanimated": "1.2.0",
+    "react-native-reanimated": "~1.7.0",
     "react-native-resource-saving-container": "^1.0.2",
-    "react-native-screens": "1.0.0-alpha.23",
+    "react-native-safe-area-context": "^0.7.3",
+    "react-native-screens": "~2.2.0",
     "react-native-simple-toast": "^1.0.0",
+    "react-native-tab-view": "^2.7.0",
     "react-native-tiny-toast": "^1.0.6",
     "react-navigation": "1.5.11",
     "react-navigation-tabs": "^2.4.1",
@@ -61,7 +69,7 @@
   "devDependencies": {
     "babel-preset-react-native": "^4.0.1",
     "flow-bin": "^0.63.1",
-    "jest-expo": "^35.0.0",
+    "jest-expo": "^37.0.0",
     "react-test-renderer": "16.0.0"
   },
   "engines": {
