@@ -9,17 +9,13 @@ import {
 import { BigButton } from '../components/BigButton';
 import { BoldText, RegularText, MediumText } from '../components/StyledText';
 import PrideraiserCampaignSummaryAdmin from '../components/PrideraiserCampaignSummaryAdmin';
-import { Ionicons } from '@expo/vector-icons';
-import { DefaultColors, Skin, Settings } from '../config/Settings';
+import { MaterialCommunityIcons } from '@expo/vector-icons';
+import { DefaultColors, Skin, Settings } from '../../config';
 import { getCampaign } from '../services/prideraiserService';
 import moment from 'moment';
-import i18n from "../../i18n";
+import i18n from '../i18n';
 
 export default class PostAttachmentComposePrideraiserMatch extends React.Component {
-    static navigationOptions = {
-        header: null
-    };
-
     state = {
         loadedCampaign: false,
         campaign: {
@@ -31,7 +27,11 @@ export default class PostAttachmentComposePrideraiserMatch extends React.Compone
         source: ""
     }
 
-    componentWillMount = async () => {
+    componentDidMount = async () => {
+        this.props.navigation.setOptions({
+            header: null
+        })
+
         try {
             const campaign = await getCampaign(Settings.Prideraiser_CampaignId)
 
@@ -102,10 +102,10 @@ export default class PostAttachmentComposePrideraiserMatch extends React.Compone
                     disabled={!this.state.loadedCampaign}
                     buttonStyle={!this.state.loadedCampaign ? { backgroundColor: "gray", marginBottom: 15 } : { marginBottom: 15 }}
                     label={i18n.t('screens.postattachmentcomposeprideraisermatch.attach')}
-                    iconName="md-add" iconPosition="right"
+                    iconName="plus" iconPosition="right"
                     onPress={() => {
-                        if (this.props.screenProps.onAttachmentComplete)
-                            this.props.screenProps.onAttachmentComplete(
+                        if (this.props.route.params.onAttachmentComplete)
+                            this.props.route.params.onAttachmentComplete(
                                 {
                                     attachmentType: "prideraisermatch",
                                     data: {
