@@ -7,11 +7,11 @@ import {
     View
 } from 'react-native';
 import { BoldText, RegularText, MediumText } from '../components/StyledText';
-import { Ionicons } from '@expo/vector-icons';
+import { MaterialCommunityIcons } from '@expo/vector-icons';
 import FadeIn from 'react-native-fade-in-image';
-import { Skin, DefaultColors, Palette } from '../config/Settings';
+import { Skin, DefaultColors, Palette } from '../../config';
 import containerStyle from './PostAttachmentContainerStyle';
-import i18n from "../../i18n";
+import i18n from '../i18n';
 
 export default class PostAttachmentPlayer extends React.Component {
     render() {
@@ -30,11 +30,11 @@ export default class PostAttachmentPlayer extends React.Component {
                     style={{ alignContent: 'center' }}
                     key={"IG: " + player.instagram}
                     onPress={() => { Linking.openURL('https://instagram.com/' + player.instagram) }}>
-                    <Ionicons
-                        name={'logo-instagram'}
+                    <MaterialCommunityIcons
+                        name={'instagram'}
                         size={30}
                         style={{
-                            color: Palette.Sky,
+                            color: DefaultColors.Secondary,
                             marginVertical: 3,
                             marginHorizontal: 5,
                             backgroundColor: 'transparent'
@@ -49,11 +49,11 @@ export default class PostAttachmentPlayer extends React.Component {
                     style={{ alignContent: 'center' }}
                     key={"Twitter: " + player.twitter}
                     onPress={() => { Linking.openURL('https://twitter.com/intent/tweet?text=@' + player.twitter + '+') }}>
-                    <Ionicons
-                        name={'logo-twitter'}
+                    <MaterialCommunityIcons
+                        name={'twitter'}
                         size={30}
                         style={{
-                            color: Palette.Sky,
+                            color: DefaultColors.Secondary,
                             marginVertical: 3,
                             marginHorizontal: 5,
                             backgroundColor: 'transparent'
