@@ -9,40 +9,40 @@ import {
 } from 'react-native';
 import { BigButton } from '../components/BigButton';
 import { BoldText, RegularText, MediumText } from '../components/StyledText';
-import { Ionicons } from '@expo/vector-icons';
-import { Skin, DefaultColors, Palette } from '../config/Settings';
-import i18n from "../../i18n";
+import { MaterialCommunityIcons } from '@expo/vector-icons';
+import { Skin, DefaultColors, Palette } from '../../config';
+import i18n from '../i18n';
 
 export default class PostAttachmentComposeSong extends React.Component {
-    static navigationOptions = {
-        header: null
-    };
-    
     state = {
         title: "",
         lyrics: ""
     }
 
     render() {
+        this.props.navigation.setOptions({
+            header: null
+        })
+
         return (
             <View style={{ flex: 1 }}>
                 <KeyboardAvoidingView behavior="height" style={styles.container}>
-                    <BoldText style={{ textAlign: 'center' }}>{i18n.t('screens.postattchmentcomposesong.composesong')}</BoldText>
+                    <BoldText style={{ textAlign: 'center' }}>{i18n.t('screens.postattachmentcomposesong.composesong')}</BoldText>
                     <TextInput
                         style={styles.titleInput}
-                        placeholder={i18n.t('screens.postattchmentcomposesong.title')}
+                        placeholder={i18n.t('screens.postattachmentcomposesong.title')}
                         onChangeText={(text) => this.setState({ title: text })} />
                     <TextInput
                         style={styles.lyricsInput}
                         multiline={true}
-                        placeholder={i18n.t('screens.postattchmentcomposesong.lyrics')}
+                        placeholder={i18n.t('screens.postattachmentcomposesong.lyrics')}
                         onChangeText={(text) => this.setState({ lyrics: text })} />
                     <BigButton
-                        label={i18n.t('screens.postattchmentcomposesong.attach')}
-                        iconName="md-add" iconPosition="right"
+                        label={i18n.t('screens.postattachmentcomposesong.attach')}
+                        iconName="plus" iconPosition="right"
                         onPress={() => {
-                            if (this.props.screenProps.onAttachmentComplete)
-                                this.props.screenProps.onAttachmentComplete(
+                            if (this.props.route.params.onAttachmentComplete)
+                                this.props.route.params.onAttachmentComplete(
                                     {
                                         attachmentType: "song",
                                         data: {
