@@ -2,7 +2,7 @@ import React from 'react';
 import { Animated, Image, FlatList, Platform, ScrollView, SectionList, StyleSheet, View } from 'react-native';
 import FadeIn from 'react-native-fade-in-image';
 import Constants from 'expo-constants';
-import { Ionicons } from '@expo/vector-icons';
+import { MaterialCommunityIcons } from '@expo/vector-icons';
 import { HeaderBackButton } from 'react-navigation';
 import { RectButton } from 'react-native-gesture-handler';
 import withUnstated from '@airship/with-unstated';
@@ -11,9 +11,9 @@ import NavigationBar from '../components/NavigationBar';
 import { RegularText, BoldText, MediumText } from '../components/StyledText';
 import LoadingPlaceholder from '../components/LoadingPlaceholder';
 import { FontSizes, Icons, Layout } from '../constants';
-import { Skin, DefaultColors, Palette } from '../config/Settings';
+import { Skin, DefaultColors, Palette } from '../../config';
 import { defaultFormatUtc } from 'moment';
-import i18n from '../../i18n';
+import i18n from '../i18n';
 
 class FoePlayerRow extends React.Component {
     render() {
@@ -48,7 +48,7 @@ class RosterFoePlayers extends React.Component {
     }
 
     componentDidMount = () => {
-        let foe = this.props.navigation.state.params.foe.item;
+        let foe = this.props.route.params.foe.item;
         let backgroundColor = DefaultColors.ButtonBackground;
         let accentColor = DefaultColors.ButtonText;
         let textColor = DefaultColors.ButtonText;
@@ -86,7 +86,7 @@ class RosterFoePlayers extends React.Component {
 
         let foeImage =
             <View>
-                <Ionicons name={Skin.Roster_FoesTabIcon} color={DefaultColors.ButtonBackground} size={150} />
+                <MaterialCommunityIcons name={Skin.Roster_FoesTabIcon} color={DefaultColors.ButtonBackground} size={150} />
                 <BoldText style={{ color: DefaultColors.ColorText, marginBottom: 10 }}>{foe.opponent}</BoldText>
             </View>
         if (foe.logo) {
