@@ -1,12 +1,8 @@
 import About from './About';
-import CapoHome from './CapoHome';
-import CapoLogin from './CapoLogin';
+import AdminHome from './AdminHome';
+import AdminLogin from './AdminLogin';
 import Channel from './Channel';
-import Events from './Events';
-import FoesSelector from './FoesSelector';
-import FoesRoster from './FoesRoster';
 import Home from './Home';
-import Instrumentation from './Instrumentation';
 import Player from './Player';
 import PostAttach from './PostAttach';
 import PostAttachmentComposeSong from './PostAttachmentComposeSong';
@@ -19,31 +15,26 @@ import PostAttachmentSelectSong from './PostAttachmentSelectSong';
 import PostCreate from './PostCreate';
 import PostPreview from  './PostPreview';
 import RefereeCard from './RefereeCard';
-import Roster from './Roster';
 import RosterFoePlayers from './RosterFoePlayers';
 import RosterFoes from './RosterFoes';
-import RosterHome from './RosterHome';
-import Shop from './Shop';
+import RosterTeam from './RosterTeam';
 import SinglePost from './SinglePost';
 import SingleSong from './SingleSong';
 import Smoke from "./Smoke";
 import Songbook from './Songbook';
-import Songs from './Songs';
-import Standings from './Standings';
+import SongbookContents from './SongbookContents';
+import SongbookCover from './SongbookCover';
+import SongbookPages from './SongbookPages';
 import TableOfContentsInline from './TableOfContentsInline';
 import TwitterList from './TwitterList';
-import Volunteer from './Volunteer';
+import WebView from './WebView';
 
 export default {
   About,
-  CapoHome,
-  CapoLogin,
+  AdminHome,
+  AdminLogin,
   Channel,
-  Events,
-  FoesSelector,
-  FoesRoster,
   Home,
-  Instrumentation,
   Player,
   PostAttach,
   PostAttachmentComposeSong,
@@ -56,18 +47,17 @@ export default {
   PostCreate,
   PostPreview,
   RefereeCard,
-  Roster,
   RosterFoePlayers,
   RosterFoes,
-  RosterHome,
-  Shop,
+  RosterTeam,
   SinglePost,
   SingleSong,
   Smoke,
   Songbook,
-  Songs,
-  Standings,
+  SongbookContents,
+  SongbookCover,
+  SongbookPages,
   TableOfContentsInline,
   TwitterList,
-  Volunteer
+  WebView
 };
