@@ -1,4 +1,4 @@
-import { HOOLIGAN_HYMNAL_SERVER_ADDRESS } from '../config/Settings';
+import { Urls } from '../../config';
 import ApiError from './ApiError';
 import React from 'react';
 
@@ -7,7 +7,7 @@ export default class ApiClient {
   }
 
   async doRequest(resource, options = {}) {
-    let url = `${HOOLIGAN_HYMNAL_SERVER_ADDRESS}/api/${resource}`;
+    let url = `${Urls.HooliganHymnalServer}/api/${resource}`;
     let response = await fetch(url, options);
     if (!response.ok) {
       throw new ApiError(`Request to ${url} returned error: ${response.status} ${response.statusText}`);
